pyproject.toml
[tool.poetry]
name = "drf-hw"
version = "0.1.0"
description = ""
authors = ["Александра <aleksandrasilina@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
black = "^24.8.0"
isort = "^5.13.2"
django-filter = "^24.3"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
